trilog/requirements.txt
# TriLog Dependencies

# OpenTelemetry Core
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation>=0.41b0

# ClickHouse
clickhouse-driver>=0.2.6
clickhouse-connect>=0.6.0

# Data Validation
pydantic>=2.0.0
jsonschema>=4.19.0

# Async Support
asyncio>=3.4.3
aiohttp>=3.8.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
structlog>=23.1.0
rich>=13.5.0

# CLI Tools
click>=8.1.0
tabulate>=0.9.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
faker>=19.0.0

# Development
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
ruff>=0.0.285
